Them xoa sua san pham

+++ admin/modules/category/delete.php
@@ -8,13 +8,21 @@
             $_SESSION['error'] = "Không tồn tại sản phẩm";
             redirectAdmin('category');
         }
-        $num = $db ->delete("category",$id);
-        if($num >0){
-            $_SESSION['success'] = ' Xóa thành công ';
-            redirectAdmin("category");
+        //kiem tra danh muc xem co san pham chua
+        $is_product = $db -> fetchOne('product'," category_id = ".$id." ");
+        if($is_product==NULL){
+            $num = $db ->delete("category",$id);
+            if($num >0){
+                $_SESSION['success'] = ' Xóa thành công ';
+                redirectAdmin("category");
+            }
+            else {
+                $_SESSION['error'] = ' Xóa thất bại ';
+                redirectAdmin("category");
+            }
         }
-        else {
-            $_SESSION['error'] = ' Xóa thất bại ';
+        else{
+            $_SESSION['error'] = ' Danh mục có sản phẩm. Bạn không thể xóa ';
             redirectAdmin("category");
         }
 ?>
